v6: major ufo upgrades and new structure
app role env layering
aws client retry improvements
CI build: acceptance test with cody
default to bridge mode instead of awsvpc
delete docs, move to separate repo
docker vars state app scope support
ecs service autoscaling support
iam role lookup path support
improve CLI help menu
log docker output in quiet mode for `ufo ship`
new commands: exec, central
refactor: remove context in favor of vars
remove commands: ships network task tasks upgrade
remove settings
ufo central concept
ufo config.rb concept
use default vpc when not configured

+++ lib/ufo/docker/cleaner.rb
@@ -1,26 +1,28 @@
-module Ufo
-  class Docker::Cleaner
-    include Util
+module Ufo::Docker
+  class Cleaner
+    include Ufo::Utils::Execute
+    include Ufo::Utils::Logging
 
     def initialize(docker_image_name, options)
       # docker_image_name does not containg the tag
       # Example: 123456789.dkr.ecr.us-east-1.amazonaws.com/image
       @docker_image_name = docker_image_name
       @options = options
-      @keep = options[:keep] || settings[:clean_keep] || 3
+      @keep = Ufo.config.docker.clean_keep
       @tag_prefix = options[:tag_prefix] || "ufo"
     end
 
     def cleanup
+      return if @keep.nil?
       return if delete_list.empty?
       command = "docker rmi #{delete_list}"
-      say "Cleaning up docker images...".color(:green)
-      say "=> #{"docker rmi #{delete_list}"}".color(:green)
+      logger.info "Cleaning docker images"
       return if @options[:noop]
-      execute(command, use_system: false) # to use_system: false silence output
+      execute(command, quiet: true)
     end
 
     def delete_list
+      return if @keep.nil?
       return [] if ENV['TEST'] || @options[:noop]
       return @delete_list if @delete_list
 
@@ -43,11 +45,6 @@ module Ufo
       else
         @delete_list = delete_tags.map { |t| "#{@docker_image_name}:#{t}" }.join(' ')
       end
-    end
-
-    def say(msg)
-      msg = "NOOP #{msg}" if @options[:noop]
-      puts msg unless @options[:mute]
     end
   end
 end
